Expose envelope API for default community
The default community API doesn't have the /:envelope_community prefix,
but otherwise mirrors the regular API exactly.

Notable change: getting info about a community now requires a /community
prefix so we don't interfere with the existing / and /info API endpoints
on the API root.
